More useful test of allPathsTo
We take the set of all nodes reachable from the start and
intersect it with the set of all nodes that can reach the end.

Then, for every node in the original graph, we should be able to build
a path from the start and a path to the end iff it's in the above set.